Hit Final Wall
Without direct access to /dev/, ffmpeg cannot find the relevant video device on linux in order to encode video as VP8. SELinux, as it has been configured on stock Android 9 (Moto G6 here, YMMV) prevents this. This app is in as a working state as I can get it on "legitimate" software, let me know if it works on a rooted install.